samples/python/agents/ai-creative-studio/agents/brand_strategist/pyproject.toml
[project]
name = "brand-strategist"
version = "0.1.0"
description = "Brand Strategist A2A Agent for market research and trend analysis"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.51.0",
    "google-adk[a2a]==1.20.0",
    "uvicorn[standard]>=0.25.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Single-file agent - include root directory
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.black]
line-length = 100
target-version = ["py311"]
